type casted to float on lines 30-35

+++ 2021-2022/botCode/highLevelPacman.py
@@ -27,12 +27,12 @@ class highLevelPacman(rm.ProtoModule):
         with open("botCode/weights.txt", "r") as f:
             lines = f.readlines()
             values = lines[0].split()
-            FREQUENCY = int(values[0])
-            FEAR = int(values[1])
-            PELLET_WEIGHT = int(values[2])
-            SUPER_PELLET_WEIGHT = int(values[3])
-            GHOST_WEIGHT = int(values[4])
-            FRIGHTENED_GHOST_WEIGHT = int(values[5])
+            FREQUENCY = float(values[0])
+            FEAR = float(values[1])
+            PELLET_WEIGHT = float(values[2])
+            SUPER_PELLET_WEIGHT = float(values[3])
+            GHOST_WEIGHT = float(values[4])
+            FRIGHTENED_GHOST_WEIGHT = float(values[5])
             print(FREQUENCY)
 
         super().__init__(addr, port, message_buffers, MsgType, FREQUENCY, self.subscriptions)
